商品機能の実装修正

+++ app/controllers/items_controller.rb
@@ -35,11 +35,15 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if @item.destroy
-      redirect_to root_path
-    else
-      render :show
-    end
+    if user_signed_in? && @item.user.id == current_user.id
+       if @item.destroy
+        redirect_to root_path
+       else
+          render :show 
+       end   
+     else
+      render :show  
+    end 
   end
 
   private
